requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
awswrangler==2.4.0
    # via -r requirements.in
beautifulsoup4==4.9.3
    # via redshift-connector
boto3==1.17.7
    # via
    #   awswrangler
    #   redshift-connector
botocore==1.20.7
    # via
    #   awswrangler
    #   boto3
    #   redshift-connector
    #   s3transfer
certifi==2019.11.28
    # via requests
chardet==3.0.4
    # via requests
cycler==0.10.0
    # via matplotlib
et-xmlfile==1.0.1
    # via openpyxl
idna==2.8
    # via requests
jdcal==1.4.1
    # via openpyxl
jmespath==0.9.4
    # via
    #   boto3
    #   botocore
joblib==1.0.1
    # via scikit-learn
kiwisolver==1.3.1
    # via matplotlib
lxml==4.6.2
    # via redshift-connector
matplotlib==3.3.4
    # via
    #   pythonpredictions-cobra
    #   seaborn
numpy==1.19.5
    # via
    #   awswrangler
    #   matplotlib
    #   pandas
    #   pyarrow
    #   pythonpredictions-cobra
    #   scikit-learn
    #   scipy
    #   seaborn
openpyxl==3.0.6
    # via awswrangler
pandas==1.1.5
    # via
    #   awswrangler
    #   pythonpredictions-cobra
    #   seaborn
pg8000==1.17.0
    # via awswrangler
pillow==8.1.0
    # via matplotlib
pyarrow==3.0.0
    # via awswrangler
pymysql==1.0.2
    # via awswrangler
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.1
    # via
    #   botocore
    #   matplotlib
    #   pandas
pythonpredictions-cobra==1.0.1
    # via -r requirements.in
pytz==2021.1
    # via
    #   pandas
    #   redshift-connector
redshift-connector==2.0.873
    # via awswrangler
requests==2.25.1
    # via
    #   -r requirements.in
    #   redshift-connector
s3transfer==0.3.1
    # via boto3
scikit-learn==0.24.1
    # via pythonpredictions-cobra
scipy==1.5.4
    # via
    #   pythonpredictions-cobra
    #   scikit-learn
    #   seaborn
scramp==1.2.0
    # via
    #   pg8000
    #   redshift-connector
seaborn==0.11.1
    # via pythonpredictions-cobra
six==1.14.0
    # via
    #   cycler
    #   python-dateutil
soupsieve==2.2
    # via beautifulsoup4
threadpoolctl==2.1.0
    # via scikit-learn
urllib3==1.25.7
    # via
    #   botocore
    #   requests
